webtentacle/requirements.txt
pyyaml
splunklib
splunk-sdk
xmltodict
unittest